pyproject.toml
[tool.poetry]
name = "event"
version = "0.1.0"
description = ""
authors = ["Louis Agyapong <agyapong.louis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
python-decouple = "^3.6"
mypy = "^0.942"
black = "^22.1.0"
django-extensions = "^3.1.5"
ipython = "^8.2.0"
django-debug-toolbar = "^3.2.4"
django-model-utils = "^4.2.0"
isort = "^5.10.1"
dj-database-url = "^0.5.0"
djangorestframework = "^3.13.1"
django-taggit = "^2.1.0"
Pillow = "^9.0.1"
django-versatileimagefield = "^2.2"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
